Further cleanup

+++ skfp/fingerprints/estate.py
@@ -34,9 +34,9 @@ class EStateFingerprint(FingerprintTransformer):
 
         X = np.array([FingerprintMol(x) for x in X])
         if self.variant == "bit":
-            X = X[:, 0] > 0
+            X = (X[:, 0] > 0).astype(int)
         elif self.variant == "count":
-            X = X[:, 0]
+            X = (X[:, 0]).astype(int)
         else:  # "sum" variant
             X = X[:, 1]
